Stop sticky from going outside constraint element.
This is so the sticky can be `position: fixed` up until it starts
to flow outside of a constraint element, at which point its
position is measured and it gets stuck at that point in the
container. It means a sidebar can have a sticky element so far as
the element is within the sidebar, which may not span the entire
height of the document.